Add Sum256 and Sum224 no-allocation functions (#1)
The New and Sum methods, part of the Hash stdlib interface, allocate memory on the heap. The stdlib sha256 package uses a different method, Sum256, that does not allocate memory.

This change adds two functions, Sum256 and Sum224, and optimizes the code so that calling them does not allocate any memory on the heap. The standard New+Write+Sum API allocations are also reduced.

New tests and benchmarks are also added, and the tests package is made different from the code one. The tests do not need access to code internals anyway, blackbox testing for the win!